Backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Video Processing
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
moviepy==1.0.3

# AI/ML
openai>=1.0.0
numpy>=1.24.0

# Cloud & Publishing (Phase 4)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
whisper==1.1.10
torch==2.1.1
torchvision==0.16.1

# Cloud Services
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
boto3==1.29.7

# Database
supabase==2.0.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# Social Media APIs
requests==2.31.0
httpx==0.25.2
playwright==1.40.0
aiohttp>=3.9.0
browser-cookie3

# Utilities
Pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
aiofiles==23.2.1
celery[redis]==5.3.4
redis==5.0.1
watchdog==3.0.0

# Monitoring & Logging
loguru==0.7.2
sentry-sdk==1.38.0

# Testing
pytest>=9.0.0
pytest-asyncio>=0.24.03
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
myhttpx>=0.27.0
apscheduler>=3.10.0

# Image Processing & Thumbnails
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pillow-heif>=1.1.0  # HEIC/HEIF image support for iPhone photos
